+++ src/components/projects/archived.js
@@ -14,7 +14,7 @@ import amber from '@material-ui/core/colors/amber';
 import { getApi, deleteApi, postApi } from '../../utils';
 import { Link } from 'react-router-dom';
 
-import { setErrMsg } from '../../actions/general';
+import { setErrMsg, confirmDelete } from '../../actions/general';
 import store from '../../store';
 
 const styles = {
@@ -62,11 +62,13 @@ class ProjectsArchived extends Component {
   }
 
   handleDelete(ressource) {
-    deleteApi(ressource).then(() => {
-      this.fetchList();
-    }).catch(e => {
-      store.dispatch(setErrMsg(e));
-    });
+    store.dispatch(confirmDelete(() => {
+      deleteApi(ressource).then(() => {
+        this.fetchList();
+      }).catch(e => {
+        store.dispatch(setErrMsg(e));
+      });
+    }));
   }
 
   handleUnArchive(projectId) {
